refactor AI components
