+++ Frontend/src/components/Chat/Chat.css
@@ -1,3 +1,5 @@
+/* Chat Component Styles */
+
 /* Chat Toggle Button */
 .chat-toggle-btn {
   position: fixed;
