+++ Frontend/src/components/Chat/ChatHeader.jsx
@@ -1,0 +1,37 @@
+import React from "react";
+
+const ChatHeader = ({ onClear }) => {
+  return (
+    <div className="chat-header">
+      <div className="chat-header-info">
+        <div className="chat-avatar">
+          <svg width="24" height="24" viewBox="0 0 24 24" fill="currentColor">
+            <path d="M12 2C6.48 2 2 6.48 2 12s4.48 10 10 10 10-4.48 10-10S17.52 2 12 2zm0 3c1.66 0 3 1.34 3 3s-1.34 3-3 3-3-1.34-3-3 1.34-3 3-3zm0 14.2c-2.5 0-4.71-1.28-6-3.22.03-1.99 4-3.08 6-3.08 1.99 0 5.97 1.09 6 3.08-1.29 1.94-3.5 3.22-6 3.22z" />
+          </svg>
+        </div>
+        <div className="chat-header-text">
+          <h3>Habit Coach</h3>
+          <span className="chat-status">
+            <span className="status-indicator"></span>
+            Online
+          </span>
+        </div>
+      </div>
+      <button className="chat-clear-btn" onClick={onClear} title="Clear chat">
+        <svg
+          width="18"
+          height="18"
+          viewBox="0 0 24 24"
+          fill="none"
+          stroke="currentColor"
+          strokeWidth="2"
+        >
+          <polyline points="1 4 1 10 7 10"></polyline>
+          <path d="M3.51 15a9 9 0 1 0 2.13-9.36L1 10"></path>
+        </svg>
+      </button>
+    </div>
+  );
+};
+
+export default ChatHeader;
