+++ Frontend/src/components/Chat/TypingIndicator.jsx
@@ -1,0 +1,15 @@
+import React from "react";
+
+const TypingIndicator = () => {
+  return (
+    <div className="chat-message assistant">
+      <div className="message-content typing-indicator">
+        <span></span>
+        <span></span>
+        <span></span>
+      </div>
+    </div>
+  );
+};
+
+export default TypingIndicator;
