+++ Frontend/src/pages/auth-pages/MainLayout.jsx
@@ -6,7 +6,7 @@ import { useAuth } from "../../context/useAuth";
 import { useNavigate } from "react-router-dom";
 import { useEffect } from "react";
 import Header from "../../components/Header";
-import Chat from "../../components/Chat";
+import Chat from "../../components/Chat/Chat";
 const MainLayout = () => {
   const { isAuthenticated } = useAuth();
   const navigate = useNavigate();
